+++ src/amplitude.js
@@ -27,7 +27,10 @@ var DEFAULT_OPTIONS = {
   saveEvents: true,
   sessionTimeout: 30 * 60 * 1000,
   unsentKey: 'amplitude_unsent',
-  uploadBatchSize: 100
+  uploadBatchSize: 100,
+  batchEvents: false,
+  eventUploadThreshold: 30,
+  eventUploadPeriodMillis: 30 * 1000 // 30s
 };
 var LocalStorageKeys = {
   LAST_EVENT_ID: 'amplitude_lastEventId',
@@ -76,11 +79,16 @@ Amplitude.prototype.init = function(apiKey, opt_userId, opt_config) {
       if (opt_config.includeReferrer !== undefined) {
         this.options.includeReferrer = !!opt_config.includeReferrer;
       }
+      if (opt_config.batchEvents !== undefined) {
+        this.options.batchEvents = !!opt_config.batchEvents;
+      }
       this.options.platform = opt_config.platform || this.options.platform;
       this.options.language = opt_config.language || this.options.language;
       this.options.sessionTimeout = opt_config.sessionTimeout || this.options.sessionTimeout;
       this.options.uploadBatchSize = opt_config.uploadBatchSize || this.options.uploadBatchSize;
+      this.options.eventUploadThreshold = opt_config.eventUploadThreshold || this.options.eventUploadThreshold;
       this.options.savedMaxCount = opt_config.savedMaxCount || this.options.savedMaxCount;
+      this.options.eventUploadPeriodMillis = opt_config.eventUploadPeriodMillis || this.options.eventUploadPeriodMillis;
     }
 
     Cookie.options({
@@ -110,9 +118,8 @@ Amplitude.prototype.init = function(apiKey, opt_userId, opt_config) {
         }
       }
     }
-    if (this._unsentEvents.length > 0) {
-      this.sendEvents();
-    }
+
+    this._sendEventsIfReady();
 
     if (this.options.includeUtm) {
       this._initUtmData();
@@ -141,6 +148,23 @@ Amplitude.prototype.isNewSession = function() {
 Amplitude.prototype.nextEventId = function() {
   this._eventId++;
   return this._eventId;
+};
+
+Amplitude.prototype._sendEventsIfReady = function() {
+  if (this._unsentEvents.length === 0) {
+    return;
+  }
+
+  if (!this.options.batchEvents) {
+    this.sendEvents();
+    return;
+  }
+
+  if (this._unsentEvents.length >= this.options.eventUploadThreshold) {
+    this.sendEvents();
+  } else {
+    setTimeout(this.sendEvents.bind(this), this.options.eventUploadPeriodMillis);
+  }
 };
 
 var _loadCookieData = function(scope) {
@@ -365,7 +389,7 @@ Amplitude.prototype._logEvent = function(eventType, eventProperties, apiProperti
       this.saveEvents();
     }
 
-    this.sendEvents();
+    this._sendEventsIfReady();
 
     return eventId;
   } catch (e) {
@@ -412,7 +436,7 @@ Amplitude.prototype.removeEvents = function (maxEventId) {
 };
 
 Amplitude.prototype.sendEvents = function() {
-  if (!this._sending && !this.options.optOut) {
+  if (!this._sending && !this.options.optOut && this._unsentEvents.length > 0) {
     this._sending = true;
     var url = ('https:' === window.location.protocol ? 'https' : 'http') + '://' +
         this.options.apiEndpoint + '/';
@@ -445,9 +469,8 @@ Amplitude.prototype.sendEvents = function() {
           }
 
           // Send more events if any queued during previous send.
-          if (scope._unsentEvents.length > 0) {
-            scope.sendEvents();
-          }
+          scope._sendEventsIfReady();
+
         } else if (status === 413) {
           //log('request too large');
           // Can't even get this one massive event through. Drop it.